+++ tests/test_auth_backends.py
@@ -162,3 +162,12 @@ class TestOAuth2Middleware(BaseTest):
         response = m(request)
         self.assertIn("Vary", response)
         self.assertIn("Authorization", response["Vary"])
+
+    def test_get_default_redirect_uri_empty_redirect_uri_raise_exception(self):
+        self.application.authorization_grant_type = Application.GRANT_AUTHORIZATION_CODE
+        self.application.redirect_uris = None
+
+        with self.assertRaises(Exception):
+            self.validator.get_default_redirect_uri(
+                client_id=self.application.client_id, request=self.request
+            )